+++ packages/dara-components/dara/components/common/radio_group.py
@@ -18,12 +18,18 @@ limitations under the License.
 from typing import Any, List, Optional, Union
 
 from pydantic import field_validator
+from typing_extensions import TypedDict
 
 from dara.components.common.base_component import FormComponent
-from dara.components.common.utils import Item
 from dara.core.base_definitions import Action
+from dara.core.definitions import ComponentInstance
 from dara.core.interactivity import NonDataVariable, UrlVariable, Variable
 from dara.core.visual.components.types import Direction
+
+
+class RadioItem(TypedDict):
+    value: Any
+    label: Union[str, ComponentInstance]
 
 
 class RadioGroup(FormComponent):
@@ -37,7 +43,7 @@ class RadioGroup(FormComponent):
 
     ```python
     from dara.core import Variable
-    from dara.components.common import RadioGroup, Item
+    from dara.components import RadioGroup, RadioItem, Text
 
     value_var_str = Variable('first')
 
@@ -49,9 +55,18 @@ class RadioGroup(FormComponent):
 
     value_var_num = Variable(1)
 
-    # or as an `Item` list
+    # or as an `RadioItem` list
     RadioGroup(
-        items=[Item(label='first',value=1), Item(label='second',value=2)],
+        items=[RadioItem(label='first',value=1), Item(label='second',value=2)],
+        value=value_var_num,
+    )
+
+    # or as an `RadioItem` list with arbitrary components
+    RadioGroup(
+        items=[
+            RadioItem(label=Text(text='first', color='red'), value=1),
+            RadioItem(label=Text(text='second', color='blue'), value=2)
+        ],
         value=value_var_num,
     )
     ```
@@ -76,7 +91,7 @@ class RadioGroup(FormComponent):
     :param id: the key to be used if this component is within a form
     """
 
-    items: Union[List[Item], NonDataVariable]
+    items: Union[List[RadioItem], List[str], NonDataVariable]
     value: Optional[Union[Variable[Any], UrlVariable[Any]]] = None
     list_styling: Optional[bool] = False
     onchange: Optional[Action] = None
@@ -85,11 +100,9 @@ class RadioGroup(FormComponent):
 
     @field_validator('items', mode='before')
     @classmethod
-    def validate_items(cls, items: Any) -> Union[List[Item], NonDataVariable]:
+    def validate_items(cls, items: Any) -> Union[List[RadioItem], NonDataVariable]:
         if isinstance(items, NonDataVariable):
             return items
-        if not isinstance(items, list):
-            raise ValueError('Items must be passed as a list to the RadioGroup component')
         if len(items) == 0:
             raise ValueError('Items list is empty, you must provide at least one item')
-        return [Item.to_item(item) for item in items]
+        return [RadioItem(value=item, label=item) if isinstance(item, str) else item for item in items]