+++ packages/dara-components/changelog.md
@@ -6,6 +6,7 @@ title: Changelog
 
 -  Fixed an issue with `CodeEditor`, `Input`, `Select`, `Textarea` components where they would bubble up Enter keypresses when confirming composition using IME input
 -  Fixed `UploadDropzone` error messages to properly distinguish between wrong file types and multiple files.
+-  Update `RadioGroup` to accept new `RadioItem` type that also accepts an arbitrary `ComponentInstance` as the label.
 
 ## 1.16.17